Create app level README
starting with some high level instructions for setting up the local evn would probably make sense. Though if me and you are the only ones working on this it might now make sense. Or maybe it will in the case either of us update something we can give a summary here